Settings and Actions Toolbar (#170)
* First Pass: Refactor settings and controls into separate toolbar component
* Fix layout padding calculation to prevent vertical overflow scrolling
* Added migration logic to remove 'actions' from saved dashboardMetrics in localStorage.
* Cleaned up Dashboard CSS.
* Refine and make explicit the layout for PerformanceDashboard.
* Make chart font sizes responsive using em-based calculations.
* Refactor DashboardForceCurve, removed any/all labels from ForceCurve.
* Standardized bottom padding to all tiles
* Maximize chart height in forceCurve
